Refactor and introduce domain package
To have separate domain classes and mvp model classes
