+++ opensrp-reporting/src/main/java/org/smartregister/reporting/contract/ReportContract.java
@@ -1,8 +1,8 @@
 package org.smartregister.reporting.contract;
 
-import org.smartregister.reporting.model.IndicatorQuery;
-import org.smartregister.reporting.model.IndicatorTally;
-import org.smartregister.reporting.model.ReportIndicator;
+import org.smartregister.reporting.domain.IndicatorQuery;
+import org.smartregister.reporting.domain.IndicatorTally;
+import org.smartregister.reporting.domain.ReportIndicator;
 
 import java.util.List;
 import java.util.Map;
