+++ opensrp-reporting/src/main/java/org/smartregister/reporting/domain/PieChartSlice.java
@@ -1,4 +1,4 @@
-package org.smartregister.reporting.model;
+package org.smartregister.reporting.domain;
 
 
 public class PieChartSlice {
