+++ opensrp-reporting/src/main/java/org/smartregister/reporting/repository/IndicatorRepository.java
@@ -5,7 +5,7 @@ import android.content.ContentValues;
 import net.sqlcipher.Cursor;
 import net.sqlcipher.database.SQLiteDatabase;
 
-import org.smartregister.reporting.model.ReportIndicator;
+import org.smartregister.reporting.domain.ReportIndicator;
 import org.smartregister.repository.BaseRepository;
 import org.smartregister.repository.Repository;
