+++ opensrp-reporting/src/main/java/org/smartregister/reporting/view/IndicatorVisualisationFactory.java
@@ -3,7 +3,7 @@ package org.smartregister.reporting.view;
 import android.content.Context;
 import android.view.View;
 
-import org.smartregister.reporting.model.ReportingIndicatorVisualization;
+import org.smartregister.reporting.domain.ReportingIndicatorVisualization;
 
 public interface IndicatorVisualisationFactory {
     View getIndicatorView(ReportingIndicatorVisualization data, Context context);
