+++ opensrp-reporting/src/test/java/org/smartregister/reporting/dao/DaoTest.java
@@ -16,8 +16,8 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.smartregister.Context;
 import org.smartregister.reporting.ReportingLibrary;
-import org.smartregister.reporting.model.IndicatorQuery;
-import org.smartregister.reporting.model.ReportIndicator;
+import org.smartregister.reporting.domain.IndicatorQuery;
+import org.smartregister.reporting.domain.ReportIndicator;
 import org.smartregister.reporting.repository.DailyIndicatorCountRepository;
 import org.smartregister.reporting.repository.IndicatorQueryRepository;
 import org.smartregister.reporting.repository.IndicatorRepository;
