+++ opensrp-reporting/src/test/java/org/smartregister/reporting/repository/IndicatorQueryRepositoryTest.java
@@ -12,7 +12,7 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.smartregister.reporting.model.IndicatorQuery;
+import org.smartregister.reporting.domain.IndicatorQuery;
 import org.smartregister.repository.Repository;
 
 @RunWith(MockitoJUnitRunner.class)
