+++ sample/src/main/java/org/smartregister/sample/presenter/SamplePresenter.java
@@ -1,10 +1,10 @@
 package org.smartregister.sample.presenter;
 
 import org.smartregister.reporting.contract.ReportContract;
-import org.smartregister.reporting.model.BaseReportIndicatorsModel;
-import org.smartregister.reporting.model.IndicatorQuery;
-import org.smartregister.reporting.model.IndicatorTally;
-import org.smartregister.reporting.model.ReportIndicator;
+import org.smartregister.reporting.domain.BaseReportIndicatorsModel;
+import org.smartregister.reporting.domain.IndicatorQuery;
+import org.smartregister.reporting.domain.IndicatorTally;
+import org.smartregister.reporting.domain.ReportIndicator;
 import org.smartregister.sample.interactor.SampleInteractor;
 
 import java.lang.ref.WeakReference;
