+++ sample/src/main/java/org/smartregister/sample/utils/SampleDataDBUtil.java
@@ -2,13 +2,11 @@ package org.smartregister.sample.utils;
 
 import android.util.Log;
 
-import net.sqlcipher.database.SQLiteDatabase;
-
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.smartregister.reporting.model.IndicatorQuery;
-import org.smartregister.reporting.model.IndicatorTally;
-import org.smartregister.reporting.model.ReportIndicator;
+import org.smartregister.reporting.domain.IndicatorQuery;
+import org.smartregister.reporting.domain.IndicatorTally;
+import org.smartregister.reporting.domain.ReportIndicator;
 import org.smartregister.reporting.repository.DailyIndicatorCountRepository;
 import org.smartregister.reporting.repository.IndicatorQueryRepository;
 import org.smartregister.reporting.repository.IndicatorRepository;
